Update README.md
The Compartmental Modelling was crudely implemented in excel before I was able to convert the whole algorithm in python. Though the mathematical equation itself is correct even on the published paper, I have since then optimized the algorithm for the model by dealing with K1 getting zero or negative values.